Add files via upload
This repository consists of the CAD file of our robot's first prototype. We have tried to replicate the same robot in person when working on our robot, with a few minor changes that we made as we saw fit. This robot has a forward steering Ackerman mechanism powered by a servo motor, with a Rev Core Hex double-shaft motor driving the vehicle. We have added three ultrasonic sensors for the distance measurement, one in the front-middle of the robot, and two at 45 degrees on the left and right of the front of the robot.